Refactor: Optimize regime statistical features with VectorBT
Integrate VectorBTRollingOptimizer and UnifiedVectorizationManager for significant performance and memory improvements.
