--- src/feature_generation/categories/regime_statistical_optimized.py
@@ -1,894 +1,0 @@
-"""
-Optimized Regime Statistical Feature Generator
-
-This module provides fully optimized statistical regime features using VectorBT's
-VectorBTRollingOptimizer and UnifiedVectorizationManager for maximum performance.
-
-Key Optimizations:
-- Full VectorBTRollingOptimizer integration
-- UnifiedVectorizationManager for advanced optimization
-- Batch processing for multiple features
-- Memory-efficient chunked processing
-- GPU acceleration support
-- Advanced caching mechanisms
-"""
-
-import numpy as np
-import pandas as pd
-from typing import Any, Dict, List, Optional, Union, Tuple
-from scipy import stats
-from scipy.signal import find_peaks
-from scipy.stats import skew, kurtosis, jarque_bera
-
-from ..core.feature_generator import FeatureGenerator, FeatureResult, VectorizedFeatureGenerator, FeatureConfig, FeatureCategory
-
-# VectorBT Rolling Optimizer
-try:
-    from ..utils.vectorbt_rolling_optimizer import get_vectorbt_rolling_optimizer, VectorBTRollingOptimizer
-    ROLLING_OPTIMIZER_AVAILABLE = True
-except ImportError:
-    ROLLING_OPTIMIZER_AVAILABLE = False
-    get_vectorbt_rolling_optimizer = None
-    VectorBTRollingOptimizer = None
-
-# Unified Vectorization Manager
-try:
-    from ..utils.optimization.unified_optimizer import get_feature_optimizer, FeatureOptimizationConfig
-    UNIFIED_OPTIMIZER_AVAILABLE = True
-except ImportError:
-    UNIFIED_OPTIMIZER_AVAILABLE = False
-    get_feature_optimizer = None
-    FeatureOptimizationConfig = None
-
-# VectorBT Optimization Mixin
-try:
-    from ..core.vectorbt_optimization_mixin import VectorBTOptimizationMixin
-    VECTORBT_MIXIN_AVAILABLE = True
-except ImportError:
-    VECTORBT_MIXIN_AVAILABLE = False
-    VectorBTOptimizationMixin = None
-
-# Base calculations
-from ..base_calculations import (
-    BaseCalculator,
-    BaseCalculationType,
-    BaseCalculationConfig,
-    create_base_calculator
-)
-
-# Import tprint for consistent logging
-from src.utils.tprint import tprint
-
-# VectorBT imports for native optimization
-try:
-    import vectorbt as vbt
-    from vectorbt.generic import (
-        rolling_mean, rolling_std, rolling_var, rolling_min, rolling_max, 
-        rolling_sum, rolling_apply, rolling_corr, rolling_cov,
-        rolling_skew, rolling_kurt, rolling_quantile
-    )
-    from vectorbt.generic import scale, rank, zscore, winsorize, clip, quantile
-    VECTORBT_AVAILABLE = True
-except ImportError:
-    VECTORBT_AVAILABLE = False
-    vbt = None
-    rolling_mean = None
-    rolling_std = None
-    rolling_var = None
-    rolling_min = None
-    rolling_max = None
-    rolling_sum = None
-    rolling_apply = None
-    rolling_corr = None
-    rolling_cov = None
-    rolling_skew = None
-    rolling_kurt = None
-    rolling_quantile = None
-    scale = None
-    rank = None
-    zscore = None
-    winsorize = None
-    clip = None
-    quantile = None
-
-# Optional GPU acceleration
-try:
-    import cupy as cp
-    CUPY_AVAILABLE = True
-except ImportError:
-    CUPY_AVAILABLE = False
-    cp = None
-
-
-class OptimizedRegimeStatisticalFeatureGenerator(VectorizedFeatureGenerator, VectorBTOptimizationMixin):
-    """
-    Fully optimized statistical regime feature generator using VectorBT's advanced capabilities.
-    
-    Features:
-    - VectorBTRollingOptimizer for optimized rolling operations
-    - UnifiedVectorizationManager for advanced optimization
-    - Batch processing for multiple features
-    - Memory-efficient chunked processing
-    - GPU acceleration support
-    - Advanced caching mechanisms
-    """
-    
-    def __init__(self, config: Optional[FeatureConfig] = None):
-        if config is None:
-            config = self._create_default_config()
-        
-        # Initialize VectorBT optimization mixin
-        if VECTORBT_MIXIN_AVAILABLE:
-            VectorBTOptimizationMixin.__init__(self)
-        
-        super().__init__(config, enable_matrix_ops=True, enable_vectorization_optimization=True)
-        
-        # Initialize VectorBT Rolling Optimizer
-        if ROLLING_OPTIMIZER_AVAILABLE:
-            self.rolling_optimizer = get_vectorbt_rolling_optimizer(
-                enable_gpu=self.enable_gpu,
-                enable_parallel=True,
-                memory_efficient=True,
-                chunk_size=1000
-            )
-        else:
-            self.rolling_optimizer = None
-        
-        # Initialize Unified Vectorization Manager
-        if UNIFIED_OPTIMIZER_AVAILABLE:
-            self.unified_optimizer = get_feature_optimizer()
-        else:
-            self.unified_optimizer = None
-        
-        # Performance tracking
-        self.performance_stats = {
-            'vectorbt_operations': 0,
-            'batch_operations': 0,
-            'chunked_operations': 0,
-            'gpu_operations': 0,
-            'cache_hits': 0,
-            'total_operations': 0,
-            'total_time': 0.0
-        }
-    
-    @classmethod
-    def _create_default_config(cls) -> FeatureConfig:
-        return FeatureConfig(
-            name="optimized_regime_statistical_features",
-            category=FeatureCategory.REGIME,
-            description="Fully optimized statistical regime features using VectorBT",
-            required_columns=["close"],
-            optional_columns=["high", "low", "open", "volume"],
-            default_lookback=32,
-            min_lookback=8,
-            max_lookback=128,
-            parameters={
-                "distribution_windows": [16, 48, 128],
-                "correlation_windows": [20, 60, 160],
-                "persistence_windows": [12, 30, 96],
-                "transition_windows": [8, 20, 64],
-                "batch_processing": True,
-                "chunked_processing": True,
-                "gpu_acceleration": False
-            },
-            matrix_optimized=True,
-            gpu_accelerated=False
-        )
-    
-    def _generate_feature(self, data: pd.DataFrame, **kwargs) -> pd.Series:
-        """Generate a single statistical regime feature as required by the base class."""
-        try:
-            # Generate all statistical regime features using batch processing
-            features_dict = self.generate_features_optimized(data, **kwargs)
-            
-            # Combine all features into a single series (use first feature as representative)
-            if features_dict:
-                first_feature_name = list(features_dict.keys())[0]
-                return pd.Series(features_dict[first_feature_name], index=data.index[:len(features_dict[first_feature_name])])
-            else:
-                # Return a simple statistical feature if no features generated
-                if 'close' in data.columns and len(data) > 1:
-                    if self.rolling_optimizer:
-                        stat_feature = self.rolling_optimizer.rolling_std(data['close'], window=5)
-                    else:
-                        stat_feature = data['close'].rolling(window=5).std().fillna(0)
-                    return pd.Series(stat_feature.values, index=data.index)
-                else:
-                    return pd.Series(np.zeros(len(data)), index=data.index)
-                
-        except Exception as e:
-            tprint(f"_generate_feature: Optimized statistical regime feature generation failed: {e}")
-            return pd.Series(np.zeros(len(data)), index=data.index)
-
-    def generate_features_optimized(self, data: pd.DataFrame, **kwargs) -> Dict[str, np.ndarray]:
-        """Generate statistical regime features using optimized VectorBT operations."""
-        features = {}
-        
-        try:
-            # Validate price data
-            if 'close' not in data.columns:
-                tprint("Warning: 'close' column not found in data")
-                return features
-            
-            close_prices = data['close'].values
-            if len(close_prices) < 8:
-                tprint(f"Warning: Insufficient data points: {len(close_prices)} < 8")
-                return features
-            
-            # Calculate returns for statistical analysis
-            returns = np.diff(np.log(close_prices))
-            returns_series = pd.Series(returns, index=data.index[1:])
-            
-            # Use batch processing for multiple features
-            if self.config.parameters.get("batch_processing", True) and self.rolling_optimizer:
-                features.update(self._generate_features_batch(returns_series, data))
-            else:
-                # Fallback to individual feature generation
-                features.update(self._generate_features_individual(returns_series, data))
-            
-        except Exception as e:
-            tprint(f"Error in optimized statistical feature generation: {e}")
-        
-        return features
-    
-    def _generate_features_batch(self, returns: pd.Series, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate features using VectorBT batch processing for maximum efficiency."""
-        features = {}
-        
-        if not self.rolling_optimizer:
-            return self._generate_features_individual(returns, data)
-        
-        try:
-            # Prepare batch operations for all windows
-            batch_operations = []
-            windows = self.config.parameters["distribution_windows"]
-            
-            # Add rolling operations for all windows
-            for window in windows:
-                if len(returns) >= window:
-                    batch_operations.extend([
-                        {
-                            'type': 'rolling',
-                            'name': f'returns_mean_{window}',
-                            'params': {'column': returns.name or 'returns', 'operation': 'mean', 'window': window}
-                        },
-                        {
-                            'type': 'rolling',
-                            'name': f'returns_std_{window}',
-                            'params': {'column': returns.name or 'returns', 'operation': 'std', 'window': window}
-                        },
-                        {
-                            'type': 'rolling',
-                            'name': f'returns_skew_{window}',
-                            'params': {'column': returns.name or 'returns', 'operation': 'skew', 'window': window}
-                        },
-                        {
-                            'type': 'rolling',
-                            'name': f'returns_kurt_{window}',
-                            'params': {'column': returns.name or 'returns', 'operation': 'kurt', 'window': window}
-                        }
-                    ])
-            
-            # Execute batch operations
-            if batch_operations:
-                # Create a DataFrame with returns for batch processing
-                returns_df = pd.DataFrame({returns.name or 'returns': returns})
-                batch_results = self.rolling_optimizer._vectorbt_batch_operations(returns_df, batch_operations)
-                
-                # Process results
-                for col in batch_results.columns:
-                    if col.startswith('returns_'):
-                        features[col] = batch_results[col].fillna(0).values
-                
-                self.performance_stats['batch_operations'] += 1
-            
-            # Generate additional statistical features
-            features.update(self._generate_distribution_features_optimized(returns, data))
-            features.update(self._generate_persistence_features_optimized(returns, data))
-            features.update(self._generate_correlation_features_optimized(returns, data))
-            features.update(self._generate_transition_features_optimized(returns, data))
-            features.update(self._generate_stability_features_optimized(returns, data))
-            
-        except Exception as e:
-            tprint(f"Error in batch feature generation: {e}")
-            # Fallback to individual processing
-            features.update(self._generate_features_individual(returns, data))
-        
-        return features
-    
-    def _generate_features_individual(self, returns: pd.Series, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Fallback individual feature generation."""
-        features = {}
-        
-        try:
-            features.update(self._generate_distribution_features_optimized(returns, data))
-            features.update(self._generate_persistence_features_optimized(returns, data))
-            features.update(self._generate_correlation_features_optimized(returns, data))
-            features.update(self._generate_transition_features_optimized(returns, data))
-            features.update(self._generate_stability_features_optimized(returns, data))
-        except Exception as e:
-            tprint(f"Error in individual feature generation: {e}")
-        
-        return features
-    
-    def _generate_distribution_features_optimized(self, returns: pd.Series, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate distribution shape features using optimized VectorBT operations."""
-        features = {}
-        windows = self.config.parameters["distribution_windows"]
-        
-        for window in windows:
-            if len(returns) < window:
-                continue
-            
-            try:
-                # Use VectorBT native rolling functions when available
-                if self.rolling_optimizer and VECTORBT_AVAILABLE:
-                    # Skewness using VectorBT native function
-                    skewness = self.rolling_optimizer.rolling_skew(returns, window=window)
-                    
-                    # Kurtosis using VectorBT native function
-                    kurtosis = self.rolling_optimizer.rolling_kurt(returns, window=window)
-                    
-                    # Normality test using optimized approach
-                    normality = self._calculate_normality_optimized(returns, window)
-                    
-                else:
-                    # Fallback to custom implementation
-                    skewness = self._calculate_rolling_skewness_optimized(returns, window)
-                    kurtosis = self._calculate_rolling_kurtosis_optimized(returns, window)
-                    normality = self._calculate_distribution_normality_optimized(returns, window)
-                
-                # Pad to match data length
-                data_len = len(data)
-                skewness_padded = self._pad_series(skewness, data_len, window)
-                kurtosis_padded = self._pad_series(kurtosis, data_len, window)
-                normality_padded = self._pad_series(normality, data_len, window)
-                
-                features[f'returns_skewness_{window}'] = skewness_padded
-                features[f'returns_kurtosis_{window}'] = kurtosis_padded
-                features[f'distribution_normality_{window}'] = normality_padded
-                
-                # Persistence features
-                skew_persistence = self._calculate_skewness_persistence_optimized(returns, window)
-                kurt_persistence = self._calculate_kurtosis_persistence_optimized(returns, window)
-                
-                features[f'skewness_persistence_{window}'] = self._pad_series(skew_persistence, data_len, window)
-                features[f'kurtosis_persistence_{window}'] = self._pad_series(kurt_persistence, data_len, window)
-                
-            except Exception as e:
-                tprint(f"Error in distribution features for window {window}: {e}")
-                continue
-        
-        return features
-    
-    def _generate_persistence_features_optimized(self, returns: pd.Series, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate statistical persistence features using optimized operations."""
-        features = {}
-        windows = self.config.parameters["persistence_windows"]
-        
-        for window in windows:
-            if len(returns) < window:
-                continue
-            
-            try:
-                # Statistical persistence using optimized approach
-                stat_persistence = self._calculate_statistical_persistence_optimized(returns, window)
-                dist_stability = self._calculate_distribution_stability_optimized(returns, window)
-                stat_strength = self._calculate_statistical_strength_optimized(returns, window)
-                
-                data_len = len(data)
-                features[f'statistical_persistence_{window}'] = self._pad_series(stat_persistence, data_len, window)
-                features[f'distribution_stability_{window}'] = self._pad_series(dist_stability, data_len, window)
-                features[f'statistical_strength_{window}'] = self._pad_series(stat_strength, data_len, window)
-                
-            except Exception as e:
-                tprint(f"Error in persistence features for window {window}: {e}")
-                continue
-        
-        return features
-    
-    def _generate_correlation_features_optimized(self, returns: pd.Series, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate cross-correlation features using optimized operations."""
-        features = {}
-        windows = self.config.parameters["correlation_windows"]
-        
-        for window in windows:
-            if len(returns) < window:
-                continue
-            
-            try:
-                # Autocorrelation using optimized approach
-                autocorr = self._calculate_returns_autocorrelation_optimized(returns, window)
-                corr_stability = self._calculate_correlation_stability_optimized(returns, window)
-                cross_corr = self._calculate_cross_correlation_features_optimized(returns, window)
-                
-                data_len = len(data)
-                features[f'returns_autocorr_{window}'] = self._pad_series(autocorr, data_len, window)
-                features[f'correlation_stability_{window}'] = self._pad_series(corr_stability, data_len, window)
-                features[f'cross_correlation_{window}'] = self._pad_series(cross_corr, data_len, window)
-                
-            except Exception as e:
-                tprint(f"Error in correlation features for window {window}: {e}")
-                continue
-        
-        return features
-    
-    def _generate_transition_features_optimized(self, returns: pd.Series, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate statistical transition features using optimized operations."""
-        features = {}
-        windows = self.config.parameters["transition_windows"]
-        
-        for window in windows:
-            if len(returns) < window * 2:
-                continue
-            
-            try:
-                # Statistical regime change detection
-                stat_change = self._detect_statistical_regime_changes_optimized(returns, window)
-                dist_transition = self._calculate_distribution_transition_probability_optimized(returns, window)
-                stat_momentum = self._calculate_statistical_momentum_optimized(returns, window)
-                
-                data_len = len(data)
-                features[f'statistical_regime_change_{window}'] = self._pad_series(stat_change, data_len, window * 2)
-                features[f'distribution_transition_{window}'] = self._pad_series(dist_transition, data_len, window * 2)
-                features[f'statistical_momentum_{window}'] = self._pad_series(stat_momentum, data_len, window * 2)
-                
-            except Exception as e:
-                tprint(f"Error in transition features for window {window}: {e}")
-                continue
-        
-        return features
-    
-    def _generate_stability_features_optimized(self, returns: pd.Series, data: pd.DataFrame) -> Dict[str, np.ndarray]:
-        """Generate statistical stability features using optimized operations."""
-        features = {}
-        windows = self.config.parameters["persistence_windows"]
-        
-        for window in windows:
-            if len(returns) < window:
-                continue
-            
-            try:
-                # Statistical stability features
-                stat_stability = self._calculate_statistical_stability_optimized(returns, window)
-                dist_entropy = self._calculate_distribution_entropy_optimized(returns, window)
-                stat_consistency = self._calculate_statistical_consistency_optimized(returns, window)
-                
-                data_len = len(data)
-                features[f'statistical_stability_{window}'] = self._pad_series(stat_stability, data_len, window)
-                features[f'distribution_entropy_{window}'] = self._pad_series(dist_entropy, data_len, window)
-                features[f'statistical_consistency_{window}'] = self._pad_series(stat_consistency, data_len, window)
-                
-            except Exception as e:
-                tprint(f"Error in stability features for window {window}: {e}")
-                continue
-        
-        return features
-    
-    def _calculate_rolling_skewness_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate rolling skewness using optimized VectorBT operations."""
-        if self.rolling_optimizer and VECTORBT_AVAILABLE:
-            return self.rolling_optimizer.rolling_skew(returns, window=window)
-        else:
-            # Fallback to pandas implementation
-            return returns.rolling(window=window).skew()
-    
-    def _calculate_rolling_kurtosis_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate rolling kurtosis using optimized VectorBT operations."""
-        if self.rolling_optimizer and VECTORBT_AVAILABLE:
-            return self.rolling_optimizer.rolling_kurt(returns, window=window)
-        else:
-            # Fallback to pandas implementation
-            return returns.rolling(window=window).kurt()
-    
-    def _calculate_normality_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate distribution normality using optimized approach."""
-        if self.rolling_optimizer and VECTORBT_AVAILABLE:
-            # Use VectorBT native functions
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=window)
-            
-            # Calculate skewness and kurtosis
-            skewness = self.rolling_optimizer.rolling_skew(returns, window=window)
-            kurtosis = self.rolling_optimizer.rolling_kurt(returns, window=window)
-            
-            # Simplified normality test using JB statistic approximation
-            jb_stat = (skewness ** 2 + (kurtosis ** 2) / 4) * window / 6
-            normality = np.exp(-jb_stat / 2)  # Approximate p-value
-            
-            return normality
-        else:
-            # Fallback implementation
-            return self._calculate_distribution_normality_optimized(returns, window)
-    
-    def _calculate_distribution_normality_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate distribution normality using optimized approach."""
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=window)
-        else:
-            rolling_mean = returns.rolling(window=window).mean()
-            rolling_std = returns.rolling(window=window).std()
-        
-        # Calculate skewness and kurtosis for normality test
-        centered = returns - rolling_mean
-        skewness = (centered ** 3).rolling(window=window).mean() / (rolling_std ** 3 + 1e-8)
-        kurtosis = (centered ** 4).rolling(window=window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Simplified normality test using JB statistic approximation
-        jb_stat = (skewness ** 2 + (kurtosis ** 2) / 4) * window / 6
-        normality = np.exp(-jb_stat / 2)  # Approximate p-value
-        
-        return normality.fillna(0)
-    
-    def _calculate_skewness_persistence_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate skewness persistence using optimized operations."""
-        if self.rolling_optimizer:
-            skewness = self.rolling_optimizer.rolling_skew(returns, window=window)
-        else:
-            skewness = returns.rolling(window=window).skew()
-        
-        # Calculate autocorrelation of skewness
-        skewness_shifted = skewness.shift(1)
-        skewness_autocorr = skewness.rolling(window=window//4).corr(skewness_shifted).fillna(0)
-        
-        return skewness_autocorr
-    
-    def _calculate_kurtosis_persistence_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate kurtosis persistence using optimized operations."""
-        if self.rolling_optimizer:
-            kurtosis = self.rolling_optimizer.rolling_kurt(returns, window=window)
-        else:
-            kurtosis = returns.rolling(window=window).kurt()
-        
-        # Calculate autocorrelation of kurtosis
-        kurtosis_shifted = kurtosis.shift(1)
-        kurtosis_autocorr = kurtosis.rolling(window=window//4).corr(kurtosis_shifted).fillna(0)
-        
-        return kurtosis_autocorr
-    
-    def _calculate_statistical_persistence_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate statistical regime persistence using optimized operations."""
-        # Calculate squared returns
-        squared_returns = returns ** 2
-        
-        # Calculate autocorrelation of squared returns
-        squared_returns_shifted = squared_returns.shift(1)
-        autocorr = squared_returns.rolling(window=window).corr(squared_returns_shifted).fillna(0)
-        
-        return autocorr
-    
-    def _calculate_distribution_stability_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate distribution stability using optimized operations."""
-        sub_window = max(2, window // 4)
-        
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=sub_window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=sub_window)
-        else:
-            rolling_mean = returns.rolling(window=sub_window).mean()
-            rolling_std = returns.rolling(window=sub_window).std()
-        
-        centered = returns - rolling_mean
-        skewness = (centered ** 3).rolling(window=sub_window).mean() / (rolling_std ** 3 + 1e-8)
-        kurtosis = (centered ** 4).rolling(window=sub_window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Calculate coefficient of variation for stability
-        if self.rolling_optimizer:
-            skew_cv = self.rolling_optimizer.rolling_std(skewness, window=window) / (self.rolling_optimizer.rolling_mean(skewness, window=window).abs() + 1e-8)
-            kurt_cv = self.rolling_optimizer.rolling_std(kurtosis, window=window) / (self.rolling_optimizer.rolling_mean(kurtosis, window=window).abs() + 1e-8)
-        else:
-            skew_cv = skewness.rolling(window=window).std() / (skewness.rolling(window=window).mean().abs() + 1e-8)
-            kurt_cv = kurtosis.rolling(window=window).std() / (kurtosis.rolling(window=window).mean().abs() + 1e-8)
-        
-        # Stability based on low coefficient of variation
-        stability = np.maximum(0, 1 - (skew_cv + kurt_cv) / 2)
-        
-        return stability.fillna(0)
-    
-    def _calculate_statistical_strength_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate statistical regime strength using optimized operations."""
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=window)
-        else:
-            rolling_mean = returns.rolling(window=window).mean()
-            rolling_std = returns.rolling(window=window).std()
-        
-        centered = returns - rolling_mean
-        skewness = (centered ** 3).rolling(window=window).mean() / (rolling_std ** 3 + 1e-8)
-        kurtosis = (centered ** 4).rolling(window=window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Strength based on deviation from normal distribution
-        deviation = np.abs(skewness) + np.abs(kurtosis - 3)
-        strength = np.maximum(0, 1 - deviation / 10)  # Normalize to 0-1
-        
-        return strength.fillna(0)
-    
-    def _calculate_returns_autocorrelation_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate returns autocorrelation using optimized operations."""
-        if self.rolling_optimizer:
-            # Use VectorBT optimized correlation
-            returns_shifted = returns.shift(1)
-            autocorr = self.rolling_optimizer.rolling_corr(returns, returns_shifted, window=window)
-        else:
-            # Fallback to pandas implementation
-            autocorr = returns.rolling(window=window).apply(
-                lambda x: x.autocorr(lag=1) if len(x) > 1 else 0,
-                raw=False
-            ).fillna(0)
-        
-        return autocorr
-    
-    def _calculate_correlation_stability_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate correlation stability using optimized operations."""
-        returns_shifted = returns.shift(1)
-        
-        if self.rolling_optimizer:
-            autocorr = self.rolling_optimizer.rolling_corr(returns, returns_shifted, window=window)
-            autocorr_variance = self.rolling_optimizer.rolling_std(autocorr, window=window//4)
-        else:
-            autocorr = returns.rolling(window=window).corr(returns_shifted)
-            autocorr_variance = autocorr.rolling(window=window//4).std()
-        
-        # Calculate stability as inverse of autocorrelation variance
-        stability = np.maximum(0, 1 - autocorr_variance)
-        
-        return stability.fillna(0)
-    
-    def _calculate_cross_correlation_features_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate cross-correlation features using optimized operations."""
-        returns_lagged = returns.shift(1)
-        
-        if self.rolling_optimizer:
-            cross_corr = self.rolling_optimizer.rolling_corr(returns, returns_lagged, window=window)
-        else:
-            cross_corr = returns.rolling(window=window).corr(returns_lagged)
-        
-        return cross_corr.fillna(0)
-    
-    def _detect_statistical_regime_changes_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Detect statistical regime changes using optimized operations."""
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=window)
-        else:
-            rolling_mean = returns.rolling(window=window).mean()
-            rolling_std = returns.rolling(window=window).std()
-        
-        centered = returns - rolling_mean
-        skew1 = (centered ** 3).rolling(window=window).mean() / (rolling_std ** 3 + 1e-8)
-        kurt1 = (centered ** 4).rolling(window=window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Shift to get second window
-        skew2 = skew1.shift(-window)
-        kurt2 = kurt1.shift(-window)
-        
-        # Calculate change ratios
-        skew_change = ((skew2 - skew1).abs() / (skew1.abs() + 1e-8)).fillna(0)
-        kurt_change = ((kurt2 - kurt1).abs() / (kurt1.abs() + 1e-8)).fillna(0)
-        
-        # Apply threshold (50% change)
-        changes = ((skew_change > 0.5) | (kurt_change > 0.5)).astype(int)
-        
-        return changes.fillna(0)
-    
-    def _calculate_distribution_transition_probability_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate distribution transition probability using optimized operations."""
-        sub_window = max(2, window // 2)
-        
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=sub_window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=sub_window)
-        else:
-            rolling_mean = returns.rolling(window=sub_window).mean()
-            rolling_std = returns.rolling(window=sub_window).std()
-        
-        centered = returns - rolling_mean
-        skewness = (centered ** 3).rolling(window=sub_window).mean() / (rolling_std ** 3 + 1e-8)
-        kurtosis = (centered ** 4).rolling(window=sub_window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Calculate volatility of statistical moments
-        skew_vol = skewness.rolling(window=window*2).std()
-        kurt_vol = kurtosis.rolling(window=window*2).std()
-        
-        # Transition probability based on moment volatility
-        transition_prob = np.minimum(1, (skew_vol + kurt_vol) / 2)
-        
-        return transition_prob.fillna(0)
-    
-    def _calculate_statistical_momentum_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate statistical momentum using optimized operations."""
-        sub_window = max(2, window // 4)
-        
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=sub_window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=sub_window)
-        else:
-            rolling_mean = returns.rolling(window=sub_window).mean()
-            rolling_std = returns.rolling(window=sub_window).std()
-        
-        centered = returns - rolling_mean
-        skewness = (centered ** 3).rolling(window=sub_window).mean() / (rolling_std ** 3 + 1e-8)
-        kurtosis = (centered ** 4).rolling(window=sub_window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Calculate trend in statistical moments using linear regression approximation
-        x = np.arange(len(skewness))
-        skew_trend = skewness.rolling(window=window).apply(
-            lambda x: np.polyfit(np.arange(len(x)), x.values, 1)[0] if len(x) > 1 else 0, raw=False
-        )
-        kurt_trend = kurtosis.rolling(window=window).apply(
-            lambda x: np.polyfit(np.arange(len(x)), x.values, 1)[0] if len(x) > 1 else 0, raw=False
-        )
-        
-        momentum = (skew_trend + kurt_trend) / 2
-        
-        return momentum.fillna(0)
-    
-    def _calculate_statistical_stability_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate statistical stability using optimized operations."""
-        sub_window = max(2, window // 4)
-        
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=sub_window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=sub_window)
-        else:
-            rolling_mean = returns.rolling(window=sub_window).mean()
-            rolling_std = returns.rolling(window=sub_window).std()
-        
-        centered = returns - rolling_mean
-        skewness = (centered ** 3).rolling(window=sub_window).mean() / (rolling_std ** 3 + 1e-8)
-        kurtosis = (centered ** 4).rolling(window=sub_window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Calculate coefficient of variation for stability
-        if self.rolling_optimizer:
-            skew_cv = self.rolling_optimizer.rolling_std(skewness, window=window) / (self.rolling_optimizer.rolling_mean(skewness, window=window).abs() + 1e-8)
-            kurt_cv = self.rolling_optimizer.rolling_std(kurtosis, window=window) / (self.rolling_optimizer.rolling_mean(kurtosis, window=window).abs() + 1e-8)
-        else:
-            skew_cv = skewness.rolling(window=window).std() / (skewness.rolling(window=window).mean().abs() + 1e-8)
-            kurt_cv = kurtosis.rolling(window=window).std() / (kurtosis.rolling(window=window).mean().abs() + 1e-8)
-        
-        # Stability based on low coefficient of variation
-        stability = np.maximum(0, 1 - (skew_cv + kurt_cv) / 2)
-        
-        return stability.fillna(0)
-    
-    def _calculate_distribution_entropy_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate distribution entropy using optimized operations."""
-        if self.rolling_optimizer:
-            rolling_min = self.rolling_optimizer.rolling_min(returns, window=window)
-            rolling_max = self.rolling_optimizer.rolling_max(returns, window=window)
-            rolling_var = self.rolling_optimizer.rolling_var(returns, window=window)
-        else:
-            rolling_min = returns.rolling(window=window).min()
-            rolling_max = returns.rolling(window=window).max()
-            rolling_var = returns.rolling(window=window).var()
-        
-        # Calculate entropy using variance approximation (much faster than histogram)
-        entropy_approx = np.log(rolling_var + 1e-8)
-        
-        # Normalize entropy to 0-1 range
-        entropy_normalized = entropy_approx / (entropy_approx.rolling(window=window*2).std() + 1e-8)
-        entropy_normalized = np.clip(entropy_normalized, 0, 1)
-        
-        return entropy_normalized.fillna(0)
-    
-    def _calculate_statistical_consistency_optimized(self, returns: pd.Series, window: int) -> pd.Series:
-        """Calculate statistical consistency using optimized operations."""
-        sub_window = max(2, window // 4)
-        
-        if self.rolling_optimizer:
-            rolling_mean = self.rolling_optimizer.rolling_mean(returns, window=sub_window)
-            rolling_std = self.rolling_optimizer.rolling_std(returns, window=sub_window)
-        else:
-            rolling_mean = returns.rolling(window=sub_window).mean()
-            rolling_std = returns.rolling(window=sub_window).std()
-        
-        centered = returns - rolling_mean
-        skewness = (centered ** 3).rolling(window=sub_window).mean() / (rolling_std ** 3 + 1e-8)
-        kurtosis = (centered ** 4).rolling(window=sub_window).mean() / (rolling_std ** 4 + 1e-8) - 3
-        
-        # Calculate autocorrelation of statistical moments
-        skewness_shifted = skewness.shift(1)
-        kurtosis_shifted = kurtosis.shift(1)
-        
-        if self.rolling_optimizer:
-            skew_corr = self.rolling_optimizer.rolling_corr(skewness, skewness_shifted, window=window)
-            kurt_corr = self.rolling_optimizer.rolling_corr(kurtosis, kurtosis_shifted, window=window)
-        else:
-            skew_corr = skewness.rolling(window=window).corr(skewness_shifted)
-            kurt_corr = kurtosis.rolling(window=window).corr(kurtosis_shifted)
-        
-        consistency = (skew_corr + kurt_corr) / 2
-        
-        return consistency.fillna(0)
-    
-    def _pad_series(self, series: pd.Series, target_length: int, window: int) -> np.ndarray:
-        """Pad series to match target length."""
-        if len(series) == target_length:
-            return series.fillna(0).values
-        
-        padded = np.full(target_length, np.nan)
-        valid_indices = min(len(series), target_length - window)
-        
-        if valid_indices > 0:
-            padded[window:window + valid_indices] = series[:valid_indices].values
-        
-        return np.nan_to_num(padded, nan=0.0)
-    
-    def get_performance_stats(self) -> Dict[str, Any]:
-        """Get performance statistics."""
-        stats = self.performance_stats.copy()
-        
-        if stats['total_operations'] > 0:
-            stats['vectorbt_usage_percentage'] = (
-                stats['vectorbt_operations'] / stats['total_operations'] * 100
-            )
-            stats['batch_usage_percentage'] = (
-                stats['batch_operations'] / stats['total_operations'] * 100
-            )
-            stats['chunked_usage_percentage'] = (
-                stats['chunked_operations'] / stats['total_operations'] * 100
-            )
-            stats['gpu_usage_percentage'] = (
-                stats['gpu_operations'] / stats['total_operations'] * 100
-            )
-            stats['cache_hit_rate'] = (
-                stats['cache_hits'] / (stats['cache_hits'] + stats['total_operations']) * 100
-            )
-            stats['average_operation_time'] = (
-                stats['total_time'] / stats['total_operations']
-            )
-        else:
-            stats['vectorbt_usage_percentage'] = 0
-            stats['batch_usage_percentage'] = 0
-            stats['chunked_usage_percentage'] = 0
-            stats['gpu_usage_percentage'] = 0
-            stats['cache_hit_rate'] = 0
-            stats['average_operation_time'] = 0
-        
-        return stats
-
-
-# Factory function for creating optimized generators
-def create_optimized_regime_statistical_generator(
-    distribution_windows: List[int] = [16, 48, 128],
-    correlation_windows: List[int] = [20, 60, 160],
-    persistence_windows: List[int] = [12, 30, 96],
-    transition_windows: List[int] = [8, 20, 64],
-    batch_processing: bool = True,
-    chunked_processing: bool = True,
-    gpu_acceleration: bool = False
-) -> OptimizedRegimeStatisticalFeatureGenerator:
-    """Create an optimized regime statistical feature generator with custom parameters."""
-    
-    config = FeatureConfig(
-        name="custom_optimized_regime_statistical_features",
-        category=FeatureCategory.REGIME,
-        description="Custom optimized statistical regime features using VectorBT",
-        required_columns=["close"],
-        optional_columns=["high", "low", "open", "volume"],
-        default_lookback=32,
-        min_lookback=8,
-        max_lookback=128,
-        parameters={
-            "distribution_windows": distribution_windows,
-            "correlation_windows": correlation_windows,
-            "persistence_windows": persistence_windows,
-            "transition_windows": transition_windows,
-            "batch_processing": batch_processing,
-            "chunked_processing": chunked_processing,
-            "gpu_acceleration": gpu_acceleration
-        },
-        matrix_optimized=True,
-        gpu_accelerated=gpu_acceleration
-    )
-    
-    return OptimizedRegimeStatisticalFeatureGenerator(config)
-
-
-# Export the optimized generator
-__all__ = [
-    'OptimizedRegimeStatisticalFeatureGenerator',
-    'create_optimized_regime_statistical_generator'
-]
